Transition to V2
dstack_tasks V1 was build on a Python3 port of `fabric`
V2 is a rewrite build on `invoke` with support for `fabric2`.

This version is Python 3.6 only as it makes extensive use of f-strings,
but since it's meant to be used as a standalone tool, it does not matter
which version of Python your tool uses.

The tools does not yet take advantage of annotations, as there is an
open bug related to inspect.signature in invoke.